Make ranges explicitly require their appropriate data types

+++ src/test/java/org/quiltmc/loader/impl/config/TestReflectiveConfig.java
@@ -32,7 +32,7 @@ public final class TestReflectiveConfig {
 	public final int b = 1;
 	public final int c = 2;
 
-	@FloatRange(min=0, max=10)
+	@IntegerRange(min=0, max=10)
 	public final int d = 3;
 	public final Vec3i vec = new Vec3i(100, 200, 300);
 	public final String whatever = "Riesling";
